init-mysql/requirements.txt
google-api-python-client==2.139.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
mysqlclient==2.2.4
